Introducing masking of the smearing distortion
Top of the EM-images often suffer from geometrical distortion
which we shall call smearing. It negatively affects the quality
of stitching process. By masking the distorted portions of the image
it is possible to overcome Inf values in coarse offsets. The mask
for smearing is based on cross-correlation of the affected image with
its shifted version. The maskis then used in _estimate_offset function.